==> UFC Analysis/03 Scripts/3.1-Merging_GeoJSONs.ipynb <==
# Merging GeoJSONS

I need to have both global country level and US states to continue my analysis, so i'm choosing to merge 2 GeoJSONs i have found on Kettle: https://www.kaggle.com/datasets/chapagain/country-state-geo-location

import os
import json

# path
path = r"D:\0 - Data Analytics\6 - Advanced Analytics & Dashboard Design\UFC Analysis"

# jsons paths
global_json_path = os.path.join(path, "02 Data", "Original Data", "countries.geo.json")
us_json_path = os.path.join(path, "02 Data", "Original Data", "us-states.json")

# Load the data
with open(global_json_path) as f:
    global_json = json.load(f)

with open(us_json_path) as f:
    us_json = json.load(f)

# check the files
print(type(global_json))
print(type(us_json))

# Check structure
print(global_json.keys())
print(us_json.keys())

# I will need to merge the json to get both the US states and the countries

# Create a new GeoJSON file that contains both the US states and the countries
merged_geojson = {"type": "FeatureCollection", "features": []}

# I will Add global countries except "United States"
for feature in global_json["features"]:
    if feature["properties"]["name"] != "United States":
        merged_geojson["features"].append(feature)

# Add detailed U.S. states
merged_geojson["features"].extend(us_json["features"])

# save the new json
with open(
    os.path.join(path, "02 Data", "Prepared Data", "merged_geojson.json"), "w"
) as f:
    json.dump(merged_geojson, f)
