==> UFC Analysis/03 Scripts/7.2-Transformaring_df_for_tableau.ipynb <==
# Imports
import os
import pandas as pd


# Set your local path
path = r"D:\0 - Data Analytics\6 - Advanced Analytics & Dashboard Design\UFC Analysis"

# Load the cleaned dataset
ufc_df = pd.read_csv(
    os.path.join(path, "02 Data", "Prepared Data", "UFC_cleaned_V2.csv")
)

# Create a working copy of the full dataset
ufc_tableau_df = ufc_df.copy()

# Convert Date to datetime and extract the Year
ufc_tableau_df["Date"] = pd.to_datetime(ufc_tableau_df["Date"], errors="coerce")
ufc_tableau_df["Year"] = ufc_tableau_df["Date"].dt.year

# Create a boolean flag for fights that went to decision
ufc_tableau_df["WentToDecision"] = ufc_tableau_df["Finish"].str.contains(
    "Decision", na=False
)

# Add Fight_ID based on the index (because original data doesn't have it)
ufc_tableau_df.reset_index(inplace=True)
ufc_tableau_df.rename(columns={"index": "Fight_ID"}, inplace=True)

columns_needed = [
    "Fight_ID",
    "Date",
    "Year",
    "Location",
    "Location_Cleaned",
    "Country",
    "RedFighter",
    "BlueFighter",
    "Winner",
    "RedOdds",
    "BlueOdds",
    "RedStance",
    "BlueStance",
    "RedReachCms",
    "BlueReachCms",
    "RedHeightCms",
    "BlueHeightCms",
    "RedAge",
    "BlueAge",
    "WeightClass",
    "TitleBout",
    "Gender",
    "Finish",
    "FinishDetails",
    "FinishRound",
    "TotalFightTimeSecs",
    "WentToDecision",
    "RedAvgSigStrLanded",
    "BlueAvgSigStrLanded",
    "RedAvgSigStrPct",
    "BlueAvgSigStrPct",
]

# Keep only the relevant columns and drop rows missing Date or Fight Time
ufc_tableau_df = ufc_tableau_df[columns_needed].dropna(
    subset=["Date", "TotalFightTimeSecs"]
)

# Export the new xlsx to work on tableau
ufc_tableau_df.to_excel(
    os.path.join(path, "02 Data", "Prepared Data", "UFC_tableau_V2.xlsx"), index=False
)
